pyproject.toml
[project]
name = "practical-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use==0.1.40",
    "langchain-openai>=0.3.1",
    "langgraph>=0.3.25",
    "langgraph-api>=0.0.46",
    "langgraph-cli>=0.1.89",
    "langsmith>=0.3.24",
    "python-dotenv>=1.1.0",
]
